bugfix & noise circuit

+++ src/cCore/core_interpolation.c
@@ -13,7 +13,7 @@ Interpolation circuits definitions.
 #include "core_interpolation.h"
 #endif
 
-int Add_i3Dlin( int owner, float* pointer, int size, int components, 
+int Add_i3Dlin( int owner, double* pointer, int size, int components, 
 	double xstep, double ystep, double zstep, double xmin, double ymin, double zmin, 
 	int xsize, int ysize, int zsize) {
